:building_construction: Rename(toss-blog): 컴포넌트 구조 변경
