+++ packages/ttoss-tech-blog/src/components/ArticleCard/articleCard.module.css
@@ -1,0 +1,98 @@
+.article-card {
+    cursor: pointer;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    max-width: 100%;
+    text-decoration: none;
+}
+
+.article-card:hover .thumb-img {
+    -webkit-box-shadow: rgba(0, 0, 33, 0.07) 0px 16px 22.4px 4.8px, rgba(0, 0, 33, 0.05) 0px 3.2px 16px 0px, rgba(0, 0, 33, 0.07) 0px 0px 1px 0px;
+            box-shadow: rgba(0, 0, 33, 0.07) 0px 16px 22.4px 4.8px, rgba(0, 0, 33, 0.05) 0px 3.2px 16px 0px, rgba(0, 0, 33, 0.07) 0px 0px 1px 0px;
+    -webkit-transform: translate3D(0, -3%, 0);
+        -ms-transform: translate3D(0, -3%, 0);
+            transform: translate3D(0, -3%, 0);
+}
+
+.article-card:hover .card-title {
+    color: #3182f6;
+}
+
+.thumb-img {
+    width: 240px;
+    height: 240px;
+    -o-object-fit: cover;
+       object-fit: cover;
+    border-radius: 14px;
+    margin-right: 48px;
+    -webkit-transition: 0.2s ease-in-out;
+    -o-transition: 0.2s ease-in-out;
+    transition: 0.2s ease-in-out;
+}
+
+.card-title {
+    display: inline-block;
+    margin-top: 0;
+    margin-bottom: 14px;
+    overflow: hidden;
+    font-size: 36px;
+    font-weight: 700;
+    line-height: 1.4;
+    color: #333d4b;
+    -o-text-overflow: ellipsis;
+       text-overflow: ellipsis;
+    word-break: keep-all;
+    overflow-wrap: break-word;
+    -webkit-transition: color 0.3s ease;
+    -o-transition: color 0.3s ease;
+    transition: color 0.3s ease;
+    -webkit-line-clamp: 2;
+}
+
+.subtitle {
+    display: block;
+    margin: 0;
+    margin-bottom: 12px;
+    line-height: 1.6;
+    color: #4e5968;
+    font-size: 17px;
+    -o-text-overflow: ellipsis;
+       text-overflow: ellipsis;
+    -webkit-line-clamp: 2;
+}
+
+.update-time {
+    color: #8b95a1;
+    font-size: 15px;
+    line-height: 1.6;
+}
+
+@media only screen and (max-width: 640px) {
+    .article-card {
+        -webkit-box-orient: vertical;
+        -webkit-box-direction: normal;
+            -ms-flex-direction: column;
+                flex-direction: column;
+    }
+
+    .thumb-img {
+        width: 100%;
+        max-height: 200px;
+        margin: 0 0 20px 0;
+    }
+
+    .card-title {
+        font-size: 26px;
+        margin-bottom: 10px;
+    }
+
+    .subtitle {
+        margin-bottom: 16px;
+        font-size: 16px;
+        line-height: 1.5;
+    }
+}
