+++ packages/ttoss-tech-blog/src/components/Footer/footer.module.css
@@ -1,0 +1,87 @@
+.footer {
+    background-color: #f9fafb;
+    padding: 50px 0 100px;
+}
+
+.footer-inner {
+    max-width: 1064px;
+    margin: auto;
+    padding: 0 67px;
+    -webkit-box-sizing: border-box;
+            box-sizing: border-box;
+}
+
+.site-group-list {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    padding-bottom: 50px;
+}
+
+.site-group:last-of-type {
+    -webkit-box-flex: 1;
+        -ms-flex: 1;
+            flex: 1;
+}
+
+.site-group {
+    width: 170px;
+    margin: 0;
+    padding: 0;
+    color: #6b7684;
+    list-style: none;
+    font-size: 15px;
+    line-height: 30px;
+    -webkit-box-sizing: border-box;
+            box-sizing: border-box;
+}
+
+.site-group-title {
+    padding-bottom: 5px;
+    color: #333d4b;
+    font-weight: 700;
+    font-weight: bold;
+}
+
+.site-group-item {
+    color: #6b7684;
+}
+
+.address {
+    color: #8b95a1;
+    font-size: 13px;
+    font-style: normal;
+    line-height: 20px;
+    white-space: pre-wrap; 
+}
+
+.company-name {
+    display: block;
+    color: #333d4b;
+    font-size: 15px;
+    font-weight: bold;
+    padding-bottom: 16px;
+}
+
+@media only screen and (max-width: 640px) {
+    .footer {
+        margin-top: 0;
+        padding: 60px 40px 80px;
+    }
+
+    .footer-inner {
+        padding: 0;
+    }
+    .site-group-list {
+        -webkit-box-orient:vertical;
+        -webkit-box-direction:normal;
+            -ms-flex-direction:column;
+                flex-direction:column;
+        padding-bottom: 0
+    }
+
+    .site-group {
+        width:100%;
+        padding-bottom: 40px
+    }
+}
