+++ packages/ttoss-tech-blog/src/components/Navbar/navbar.module.css
@@ -1,0 +1,174 @@
+.navbar {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    -webkit-box-pack: justify;
+        -ms-flex-pack: justify;
+            justify-content: space-between;
+    position: sticky;
+    top: 0;
+    z-index: 9;
+    width: 100%;
+    border-bottom: 1px solid rgba(0,27,55,0.1);
+    background-color: #fff;
+}
+
+.navbar-inner {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    height: 100%;
+    width: 92%;
+    margin: 0 auto;
+    max-width: calc(100% - 48px);
+}
+
+.logo-container {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    margin-right: 24px;
+    min-height: 60px;
+}
+
+.logo-container a {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+}
+
+.logo-container object {
+    display: inline-block;
+    height: auto;
+    width: auto;
+    min-width: 110px;
+    max-height: 20px;
+}
+
+.dropdown-toggler {
+    background: none;
+    outline: none;
+    border: none;
+    cursor: pointer;
+}
+
+.dropdown-toggler .icon {
+    display: inline-block;
+    height: 24px;
+    width: 24px;
+    min-width: 24px;
+    color: rgb(176, 184, 193);
+}
+
+.dropdown-toggler .icon svg {
+    fill: currentColor;
+    color: rgb(176, 184, 193);
+    width: 100%;
+    height: 100%;
+}
+
+.right-on-mobile {
+    display: none;
+}
+
+.right-on-mobile button {
+    cursor: pointer;
+    border: 0;
+    padding: 0;
+    outline: none;
+    background: none;
+}
+
+.right-on-mobile .icon {
+    height: 24px;
+    width: 24px;
+    min-width: 24px;
+    color: rgb(176, 184, 193);
+}
+
+.navbar-content {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-flex: 1;
+        -ms-flex: 1;
+            flex: 1;
+    max-height: 100%;
+    -webkit-box-pack: justify;
+        -ms-flex-pack: justify;
+            justify-content: space-between;
+}
+
+.navbar-menu {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-wrap: wrap;
+        flex-wrap: wrap;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    padding: 0;
+    margin: 0;
+    margin-left: auto;
+}
+
+.navbar-item {
+    cursor: pointer;
+    color: #4e5968;
+    font-size: 15px;
+    line-height: 20px;
+    white-space: nowrap;
+    border: 0;
+    border-radius: 8px;
+    background-color: transparent;
+    text-align: left;
+    text-decoration: none;
+    margin: 0 8px;
+    padding: 12px 18px;
+}
+
+.navbar-item:hover {
+    background-color: rgba(2,32,71,0.05);
+}
+
+@media (max-width: 639px) {
+    .navbar-inner {
+        -webkit-box-orient: vertical;
+        -webkit-box-direction: normal;
+            -ms-flex-direction: column;
+                flex-direction: column;
+        -webkit-box-align: start;
+            -ms-flex-align: start;
+                align-items: flex-start;
+    }
+
+    .navbar-menu {
+        margin-left: 0;
+    }
+
+    .navbar-item {
+        width: 100%;
+        margin: 0;
+        padding: 18px 20px;
+        border-radius: 0;
+    }
+
+    .navbar-content {
+        width: calc(100% + 48px);
+        margin: 0 -24px;
+        -webkit-box-orient: vertical;
+        -webkit-box-direction: normal;
+            -ms-flex-direction: column;
+                flex-direction: column;
+    }
+}
