+++ packages/ttoss-tech-blog/src/components/constants.js
@@ -1,12 +1,42 @@
 export const LOGO_ARIA_LABEL = '토스 기술 블로그, 토스테크';
 
-export const LIST_LABEL = {
-  tech: '개발',
-  design: '디자인',
-};
-
 export const COPYRIGHT = 'Copyright © Viva Republica, Inc. All Rights Reserved.';
 
 export const COMPANY_NAME = '㈜비바리퍼블리카';
 
 export const EMOJI_PROFILE = 'https://static.toss.im/3d-emojis/u1F600.png';
+
+export const SITE_GROUP_LISTS = [
+  {
+    title: '토스테크',
+    items: ['의견 보내기'],
+  },
+  {
+    title: '토스',
+    items: ['홈페이지', '회사 소개', '채용'],
+  },
+  {
+    title: '고객센터',
+    items: [
+      '전화: 1599-4905 (24시간 연중무휴)',
+      '이메일: support@toss.im',
+      '카카오톡: @toss',
+    ],
+  },
+];
+
+export const LIST_LABEL = {
+  tech: '개발',
+  design: '디자인',
+};
+
+export const NAV_LIST = [
+  {
+    title: '디자인',
+    href: '/design',
+  },
+  {
+    title: '개발',
+    href: '/tech',
+  },
+];
